backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "amoneyplan"
version = "0.1.0"
description = "A personal money management app"
readme = "README.md"
authors = ["Martin Ronquillo"]
packages = [{include = "src"}]
repository = "https://github.com/RonquilloAeon/amoneyplan"

[tool.poetry.dependencies]
python = "^3.13"
strawberry-graphql = "^0.261.1"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
django-cors-headers = "^4.7.0"
pre-commit = "^4.1.0"
uvicorn = "^0.34.0"
django-allauth = "^65.5.0"
requests = "^2.32.3"
pyjwt = "^2.10.1"
cryptography = "^44.0.2"
cuid2 = "^2.0.1"
alt-profanity-check = "^1.6.1"
strawberry-graphql-django = "^0.57.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
ruff = "^0.9.9"
nox = "^2025.2.9"
pytest-django = "^4.10.0"
pytest-cov = "^4.1.0"
mypy = "^1.9.0"
faker = "^37.1.0"
pytest-profiling = "^1.8.1"

[tool.pytest.ini_options]
addopts = "-s"
testpaths = "tests"
python_files = "test_*.py"
django_find_project = true

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
]
show_missing = true
fail_under = 80

[tool.ruff]
line-length = 110
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.pycodestyle]
ignore-overlong-task-comments = true

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
